+++ R/bundleFiles.R
@@ -24,6 +24,8 @@
 #' option `rsconnect.max.bundle.files`). This prevents you from accidentally
 #' bundling a very large direcfory (i.e. you home directory).
 #'
+#' Supported servers: All servers
+#'
 #' @inheritParams deployApp
 #' @param error_call The call or environment for error reporting; expert
 #'   use only.
@@ -228,15 +230,17 @@ isPythonEnv <- function(dir, files) {
 }
 
 enforceBundleLimits <- function(appDir, totalFiles, totalSize) {
-  maxSize <- getOption("rsconnect.max.bundle.size", 3000 * 1024^2)
-  maxFiles <- getOption("rsconnect.max.bundle.files", 10000)
+  maxSize <- getOption("rsconnect.max.bundle.size", defaultMaxBundleSize)
+  maxFiles <- getOption("rsconnect.max.bundle.files", defaultMaxBundleFiles)
 
   if (totalSize > maxSize) {
     cli::cli_abort(c(
       "{.arg appDir} ({.path {appDir}}) is too large to be deployed.",
       x = "The maximum size is {maxSize} bytes.",
       x = "This directory is at least {totalSize} bytes.",
-      i = "Remove some files or adjust the rsconnect.max.bundle.size option."
+      i = "Remove some files or adjust the rsconnect.max.bundle.size option.",
+      " " = "e.g., options(rsconnect.max.bundle.size = 6 * 1024^3)",
+      " " = "See {.topic rsconnect::rsconnectOptions} for additional guidance."
     ))
   }
 
@@ -245,7 +249,9 @@ enforceBundleLimits <- function(appDir, totalFiles, totalSize) {
       "{.arg appDir} ({.path {appDir}}) is too large to be deployed.",
       x = "The maximum number of files is {maxFiles}.",
       x = "This directory contains at least {totalFiles} files.",
-      i = "Remove some files or adjust the rsconnect.max.bundle.files option."
+      i = "Remove some files or adjust the rsconnect.max.bundle.files option.",
+      " " = "e.g., options(rsconnect.max.bundle.files = 15000)",
+      " " = "See {.topic rsconnect::rsconnectOptions} for additional guidance."
     ))
   }
 }