+++ tests/testthat/test-client.R
@@ -1,13 +1,13 @@
-test_that("cloud accounts create cloud clients", {
-  account <- list(server = "posit.cloud")
-  client <- clientForAccount(account)
-  expect_equal(client$service(), "posit.cloud")
-})
-
 test_that("shinyapps accounts create shinyapps clients", {
   account <- list(server = "shinyapps.io")
   client <- clientForAccount(account)
   expect_equal(client$service(), "shinyapps.io")
+})
+
+test_that("connect cloud accounts create connect cloud clients", {
+  account <- list(server = "connect.posit.cloud")
+  client <- clientForAccount(account)
+  expect_equal(client$service(), "connect.posit.cloud")
 })
 
 test_that("connect accounts create connect clients", {