+++ NEWS.md
@@ -1,3 +1,33 @@
+# rsconnect 1.6.0
+
+* Support deploying to Posit Connect Cloud. Use `connectCloudUser()` to add
+  Connect Cloud credentials.
+
+* `rsconnect` now sets the `rsconnect.max.bundle.size` and
+  `rsconnect.max.bundle.files` options to their default values on startup
+  if they have not yet been set. (#1204)
+
+* The default `rsconnect.max.bundle.size` limit has increased to 5 GiB. (#1200)
+
+* `getLogs()` returns log lines for a shinyapps.io hosted application. (#1209)
+
+* Python environment inspection errors include the path to the target Python
+  binary. (#1207)
+
+* Improve cookie expiration date handling. (#1212)
+
+* Improve documentation and advice for `deployApp(envVars...)`.
+
+* Removed support for publishing to Posit Cloud. (#1215)
+
+  Existing Posit Cloud account records may be removed by using
+  `removeAccount("yourname", "posit.cloud")`.
+  
+  Existing Posit Cloud deployment records may be removed by using
+  `forgetDeployment(name="deployment", account="yourname", server="posit.cloud")`.
+
+* Removed the Posit Cloud-exclusive `space` argument from `deployApp()`. (#1215)
+
 # rsconnect 1.5.1
 
 * Address user registration for Posit Connect deployments hosted in Snowpark